% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenvalues.R
\name{eigenvalues_view}
\alias{eigenvalues_view}
\title{六月特征值按年月查询}
\usage{
eigenvalues_view(token, year, month)
}
\arguments{
\item{month}{}
}
\value{

}
\description{
六月特征值按年月查询
}
\examples{
eigenvalues_view()
}
