% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NationalMacroData.R
\name{NationalMacroData_view}
\alias{NationalMacroData_view}
\title{国家宏观数据按年月查询}
\usage{
NationalMacroData_view(token, year, month)
}
\arguments{
\item{month}{}
}
\value{

}
\description{
国家宏观数据按年月查询
}
\examples{
NationalMacroData_view()
}
