% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnOrder.R
\name{returnOrder_view}
\alias{returnOrder_view}
\title{退货单按年月查询}
\usage{
returnOrder_view(token, year, month)
}
\arguments{
\item{month}{}
}
\value{

}
\description{
退货单按年月查询
}
\examples{
returnOrder_view()
}
