% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesOrder_shop.R
\name{salesOrder_shop_view}
\alias{salesOrder_shop_view}
\title{销售订单匹配数据按年月查询}
\usage{
salesOrder_shop_view(token, year, month)
}
\arguments{
\item{month}{}
}
\value{

}
\description{
销售订单匹配数据按年月查询
}
\examples{
salesOrder_shop_view()
}
